requirements-dev.txt
Django==4.2.16
django-currentuser
django-environ
django-adminlte-ui==2.1.1
django-multiselectfield==0.1.13
cryptography==43.0.3
pyOpenSSL==24.2.1
django-localflavor
django-picklefield
django-reversion==5.1.0
django-timezone-field
django-wkhtmltopdf
djangorestframework
jsonfield
psycopg2-binary
python-slugify
signxml
xmlschema
xmltodict
json2xml
beautifulsoup4
pandas
lxml
coreapi
django-filter
djangorestframework-simplejwt
django-rest-passwordreset
django-cors-headers
reportlab
django-allauth
whitenoise
requests
python-dateutil
dicttoxml
untangle
xlwt==1.3.0
dotmap
libesocial @ https://github.com/qualitaocupacional/libesocial/archive/main.zip#sha256=fafd215f0367ec774052dd168520a3bd45db56ddf1010f96b80ad3cbe579ae6f
django-constance==4.1.2
setuptools==65.5.1

pip-audit==2.7.3  # DEV
django-debug-toolbar  # DEV
flake8  # DEV
mypy  # DEV
types-beautifulsoup4
lxml-stubs
djangorestframework-stubs
types-python-dateutil
types-xmltodict
